refactor: Create simple makeTheme to create a theme according to its type
